Adding the VibrateCmd message
This sort of replaces the SingleMotorVibrateCmd, although
that command is still supported for backwards compatibilty: it will apply the speed to all vibrators on a device.

The much cooler changes here are in the message serializer and the base ButtplugMessage class. See the MessageVersioningVersion and MessageVersioningPrevious members: these tell you when a message was introduced into the schema and what the previous version of the message was. This allows messages that have been updated to be downgraded on the fly for older clients.