Adding support for sending the accelerometer results back to the client.
Also adding generic control messages for linear motion, rotation, constriction (this should cover all the features of the currently supported devices, athough I'm aware that the Domi has lights and the the realtouch has heaters and a lube squirter). The Vorze, onyx and launch have yet to be updated to handle the new generic commands.