This is the device driver for the QuadMouse (the internal guts of the SensorRod that was my SexTechHack 2017 project)
It adds movement messages to the schema, an option to the example client to start/stop sending that data and a serial driver for reading data from the Arduino.

The searial driver is a mess, as it started as a copy of the ET312 driver: this is not how you write a driver.